fix: fix type errors that surface during prepack (#110)
Didn't catch that `@comapeo/schema` wasn't upgraded to the compatible version in #107 , which was causing the [type errors seen when trying to publish a new version](https://github.com/digidem/comapeo-core-react/actions/runs/17587542167).

Also adds a CI check to run the prepack step for all PRs to avoid similar issues in the future.